requirements.txt
Flask==3.0.3
gunicorn==22.0.0
Werkzeug==3.0.6
PyPDF2==3.0.1
python-docx==0.8.11
spacy
sentence-transformers==3.4.1
faiss-cpu==1.9.0
numpy
py2neo
tokenizers==0.19.1 
transformers==4.43.0
